requirements.txt
llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-embeddings-huggingface
llama-index-vector-stores-qdrant
qdrant-client
streamlit
python-dotenv
pdf2image
pdfplumber
Pillow
